Created button to choose different date and added ImgInfo component
