+++ src/App.css
@@ -4,7 +4,7 @@
 
 .App-logo {
   /* animation: App-logo-spin infinite 20s linear; */
-  height: 40vmin;
+  height: 30vmin;
   pointer-events: none;
 }
 
@@ -30,4 +30,12 @@
   to {
     transform: rotate(360deg);
   }
+}
+
+button {
+  padding: 10px 20px;
+  margin: 20px;
+  border-radius: 10px;
+  background-color: #282c34;
+  font-size: 20px;
 }
