+++ src/Component/APOD/ImgInfo.js
@@ -1,0 +1,14 @@
+import React from "react";
+
+function ImgInfo(props) {
+    
+    return (
+        <div>
+            <p>{props.dateInfo}</p>
+            <p>{props.imgTitle}</p>
+            <p>{props.imgExplanation}</p>
+        </div>
+    )
+}
+
+export default ImgInfo;
